TaskUpdater and nullifier
TaskUpdater is a new class to hold methods related to modifying tasks
and documents based on data returned from users.

nullify is the simplest of the tasks: it simply removes the
property-object pair from the task section and does not add it to the
graph at all.

It basically means a move from:

G1: <s> <potentiallyIncorrect> <o>.
G2: <s> <...> <...> .

to {} (with the removal of graph G2 and an update of graph G1).