Connection: fix IllegalStateException
DefaultHttpClient handles only one connection at a time, but it is
possible to fire off two requests in very close succession that cause it
to throw an IllegalStateException.

Instead, the Connection class now creates a new DefaultHttpClient for
every HTTP request it makes.

This means that a fair amount of the class is now redundant: the
authenticationCallback isn't required (as authentication details are
handed to the DefaultHttpClient immediately after construction if
available), and the authenticating/setAuthenticating boolean property
isn't required either.

The hacky .duplicate() solution has also been removed, along with uses
of duplicate() and attempts to get the client property directly.